% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{est_ci_str}
\alias{est_ci_str}
\title{Estimate and confidence interval string maker}
\usage{
est_ci_str(est, lcb, ucb, d = 2, cid = d)
}
\arguments{
\item{est}{The point estimate.}

\item{lcb}{The lower confidence bound.}

\item{ucb}{The upper confidence bound.}

\item{d}{Digits for the estimate.}

\item{cid}{Digits for the confidence interval (defaults to d)}
}
\description{
Compresses an estimate, lower (lci) and upper (uci) confidence interval into a string.
}
\details{
cid is for confidence interval digits.  If it is omitted then d is used.  d is for the number of digits in the estimate.
}
