% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_and_string.R
\name{permute_and_string}
\alias{permute_and_string}
\title{Permute and string}
\usage{
permute_and_string(a, b, c)
}
\arguments{
\item{a}{An item}

\item{b}{Some other item}

\item{c}{A third item}
}
\description{
Randomly choose an order for a,b,c.  Return a comma-space
  separated string.
}
\details{
A completely pointless function - fitting for a pointless package!
}
